feature: calculate and change route
