+++ settlement-service/src/constants/constants.ts
@@ -1,0 +1,3 @@
+class Constants {
+    static readonly CURRENCY = 'THB'
+}
